yolov11_mlflow/requirements.txt
ultralytics>=8.0.0
mlflow>=2.8.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
boto3>=1.28.0
psycopg2-binary>=2.9.6
pydantic>=2.0.0
pytest>=7.3.1
pre-commit>=3.3.2
python-multipart>=0.0.6
pillow>=9.5.0
pandas>=2.0.0
matplotlib>=3.7.1
requests>=2.31.0
dvc>=3.30.0
dvc-s3>=2.23.0 